+++ NEWS.md
@@ -1,8 +1,10 @@
 # phenoptr 0.2.8.9000
 
-- Improved the way `read_cell_seg_data` determines column types to prevent
+- Add `add_geometry` and `read_phenochart_polygons` in support of 
+  ROI processing in phenoptrReports.
+- Improve the way `read_cell_seg_data` determines column types to prevent
   errors in some edge cases.
-  
+
 # phenoptr 0.2.8
 **2020-08-22**